FIX: 이메일 수정 관련 컴포넌트 분리
